+++ src/store/modules/connext.ts
@@ -1,6 +1,6 @@
 import { MutationTree, ActionTree, GetterTree } from 'vuex'
 import { ActiveTransaction, NxtpSdk, NxtpSdkEvents } from '@connext/nxtp-sdk'
-import { utils } from 'ethers'
+import { ethers, utils } from 'ethers'
 
 import { RootState } from '@/store'
 import { networks, isProduction } from '@/config'
@@ -49,9 +49,11 @@ const mutations = <MutationTree<ConnextState>>{
 }
 
 const actions = <ActionTree<ConnextState, RootState>>{
-  async instantiateConnext() {
+  async instantiateConnext(_, signer: any) {
     console.log('Instantiate Connext, production = ', isProduction)
-    connextSDK = await instantiateConnextSDK()
+    if (!connextSDK) {
+      connextSDK = await instantiateConnextSDK(signer)
+    }
     console.log('connext after instantiating', connextSDK)
   },
 
@@ -99,8 +101,7 @@ const actions = <ActionTree<ConnextState, RootState>>{
     }
 
     if (!sendingAsset || !receivingAsset) {
-      console.error('Sending or Receiving asset not defined')
-      return
+      throw new Error('Sending or Receiving asset not defined')
     }
 
     // get amount in decimals
@@ -141,8 +142,7 @@ const actions = <ActionTree<ConnextState, RootState>>{
 
   async prepareTransfer({ state, commit }) {
     if (!state.quote) {
-      console.error('no quote')
-      return
+      throw new Error('no quote')
     }
 
     commit(types.SET_PREPARING_SWAP, true)
@@ -167,8 +167,7 @@ const actions = <ActionTree<ConnextState, RootState>>{
 
   async fulfillTransfer({ state, commit, dispatch }) {
     if (!state.prepared) {
-      console.error('not prepared')
-      return
+      throw new Error('not prepared')
     }
     if (!connextSDK) {
       await dispatch('instantiateConnext')
@@ -186,6 +185,11 @@ const actions = <ActionTree<ConnextState, RootState>>{
     { dispatch, rootState },
     activeTransaction: ActiveTransaction
   ) {
+    if (!rootState.wallet.connected) {
+      await dispatch('connectWallet')
+    }
+    if (!connextSDK) throw new Error('Connext not instantiated')
+
     const {
       crosschainTx,
       status,
@@ -194,8 +198,7 @@ const actions = <ActionTree<ConnextState, RootState>>{
       encryptedCallData,
     } = activeTransaction
     if (!activeTransaction || !crosschainTx) {
-      console.error('Missing data, unable to fulfill Connext transfer')
-      return
+      throw new Error('Missing data, unable to fulfill Connext transfer')
     }
     const { receiving, invariant } = crosschainTx
     const receivingTxData =
@@ -207,13 +210,6 @@ const actions = <ActionTree<ConnextState, RootState>>{
         : undefined
 
     if (status === NxtpSdkEvents.ReceiverTransactionPrepared) {
-      if (!connextSDK) {
-        await dispatch('instantiateConnext')
-      }
-      if (!rootState.wallet.connected) {
-        await dispatch('connectWallet')
-      }
-
       const finish = await connextSDK.fulfillTransfer(
         {
           bidSignature,
@@ -233,17 +229,15 @@ const actions = <ActionTree<ConnextState, RootState>>{
     { dispatch, rootState },
     activeTransaction: ActiveTransaction
   ) {
+    if (!rootState.wallet.connected) {
+      await dispatch('connectWallet')
+    }
+    if (!connextSDK) throw new Error('Connext not instantiated')
+
     const { sending, invariant } = activeTransaction.crosschainTx
     const sendingTxData = {
       ...invariant,
       ...sending,
-    }
-
-    if (!connextSDK) {
-      await dispatch('instantiateConnext')
-    }
-    if (!rootState.wallet.connected) {
-      await dispatch('connectWallet')
     }
 
     await connextSDK.cancel(
@@ -255,10 +249,7 @@ const actions = <ActionTree<ConnextState, RootState>>{
 
 const getters = <GetterTree<ConnextState, RootState>>{
   getActiveConnextTxs: () => async () => {
-    if (!connextSDK) {
-      connextSDK = await instantiateConnextSDK()
-      console.log('connext after instantiating', connextSDK)
-    }
+    if (!connextSDK) return
 
     const activeTxs = await connextSDK.getActiveTransactions()
     return activeTxs.map((tx: any) => {
@@ -286,7 +277,9 @@ const getters = <GetterTree<ConnextState, RootState>>{
   },
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   getTransaction: () => async () => {
-    connextSDK = await instantiateConnextSDK()
+    if (!connextSDK) {
+      throw new Error('cannot fetch transaction, connect wallet first')
+    }
     const query = `
       {
         transactions(orderBy: preparedTimestamp, orderDirection: desc, where: { transactionId: '0xd3a053e2db95eb6ca25eeb02bd27ab99031e25800e4160b197304c2ba1957acf' }) {