+++ src/store/modules/sdk.ts
@@ -136,11 +136,10 @@ const actions = <ActionTree<SDKState, RootState>>{
     commit(types.SET_BALANCE, balance)
   },
 
-  registerSigner(_, network: NetworkMetadata) {
+  registerSigner({ rootGetters }, network: NetworkMetadata) {
     console.log('registering signer for ', network)
     const networkName = network.name
-    const provider = new providers.Web3Provider(window.ethereum)
-    const newSigner = provider.getSigner()
+    const signer = rootGetters.getSigner()
 
     nomad.clearSigners()
     nomad.missingProviders
@@ -150,7 +149,7 @@ const actions = <ActionTree<SDKState, RootState>>{
         nomad.registerRpcProvider(networkName, network.rpcUrl)
       })
 
-    nomad.registerSigner(networkName, newSigner)
+    nomad.registerSigner(networkName, signer)
   },
 
   async send(
@@ -259,32 +258,21 @@ const getters = <GetterTree<SDKState, RootState>>{
   },
   blacklist: (state: SDKState) => () => state.blacklist,
   getGasPrice: () => async (network: string | number) => {
-    try {
-      const provider = nomad.getProvider(network)
-      const gasPrice = await provider?.getGasPrice()
-      return gasPrice
-    } catch (e) {
-      console.error(e)
-    }
+    const provider = nomad.getProvider(network)
+    const gasPrice = await provider?.getGasPrice()
+    return gasPrice
   },
 
   getTxMessage:
     () =>
     async (tx: TXData): Promise<TransferMessage | undefined> => {
       const { network, hash } = tx
-      let message
 
-      try {
-        message = await nomadSDK.TransferMessage.singleFromTransactionHash(
-          nomad,
-          network,
-          hash
-        )
-      } catch (e) {
-        console.error(e)
-      }
-
-      return message
+      return await nomadSDK.TransferMessage.singleFromTransactionHash(
+        nomad,
+        network,
+        hash
+      )
     },
 
   getTimestamp: